add some more unit tests
